A lot of clean up, after consuming the tool as a library.
lots of work was required.
