+++ cmd/root.go
@@ -57,6 +57,7 @@ func init() {
     rootCmd.AddCommand(GetReportCommand())
     rootCmd.AddCommand(GetHTMLReportCommand())
     rootCmd.PersistentFlags().BoolP("top", "t", false, "Only show latest changes (last git revision against HEAD)")
+    rootCmd.PersistentFlags().IntP("limit", "l", 5, "Limit history to number of revisions (default is 5)")
 }
 
 func initConfig() {
