+++ html-report/templates/header.gohtml
@@ -23,8 +23,11 @@
         <meta name="twitter:data1" content="openapi-changes"/>
         <style>
             .preloader {
-                width: 100vh;
-                margin: calc(50vh - 611px) auto;
+                width: 400px;
+                height: 200px;
+                position: absolute;
+                top: calc(50vh - 100px);
+                left: calc((100vw - 50%) - 200px);
                 text-align: center;
             }
             .ellipsis {
