+++ html-report/ui/src/OpenAPIChanges.css
@@ -7,24 +7,17 @@
 #main_container {
   font-family: var(--font-stack);
   flex: 1 0 auto;
-  width: calc(100vw - 50px);
-}
-
-body[data-embedded] #main_container {
   width: 100vw;
 }
+
 
 
 hr {
   border: none;
   border-bottom: 1px dashed var(--secondary-color);
-}
-
-body[data-embedded] hr {
   margin-block-end: 0;
   margin-block-start: 0;
 }
-
 
 .height-100 {
   height: 100px;
@@ -99,12 +92,17 @@ body[data-embedded] hr {
   width: 100%
 }
 
-body[data-embedded] section.report-header > header {
+section.report-header > header {
   padding-left: 10px;
   padding-right: 10px;
-  padding-top: 10px;
+  padding-top: 5px;
   height: 40px;
 }
+
+body[data-embedded] section.report-header > header {
+  padding-top: 10px;
+}
+
 
 
 .breaking-change {
@@ -162,6 +160,7 @@ body[data-embedded] section.report-header > header {
 
 footer {
   font-size: 0.6em;
+  background-color: var(--background-color);
 }
 
 body[data-embedded] footer {
