+++ html-report/ui/src/OpenAPIChanges.tsx
@@ -1,8 +1,11 @@
-import React, {createContext, useContext, useEffect, useState} from 'react';
+// Copyright 2023 Princess B33f Heavy Industries / Dave Shanley
+// SPDX-License-Identifier: MIT
+
+import React, {useEffect} from 'react';
 import './variables.css';
 import "allotment/dist/style.css";
 import './OpenAPIChanges.css';
-import {Col, ConfigProvider, Row, Tabs, theme} from 'antd';
+import {ConfigProvider, Tabs, theme} from 'antd';
 import {Header} from "./components/header";
 import {ReportContainer} from "./components/report-statistics/ReportContainer";
 import {GraphViewComponent} from "@/components/graph/GraphView";
@@ -21,7 +24,7 @@ export interface OpenAPIChangesProps {
 
 export let ReportStoreContext: React.Context<StoreApi<ReportState>>
 
-function OpenAPIChanges(props: OpenAPIChangesProps) {
+export function OpenAPIChanges(props: OpenAPIChangesProps) {
     const closeDrawer = useDrawerStore((state: DrawerState) => state.closeDrawer)
     const nodeMap: Map<String, TreeGraphMap> = new Map<String, TreeGraphMap>();
     const lookupMap = useChangeStore((state: ChangeState) => state.treeMapLookup)
@@ -57,7 +60,6 @@ function OpenAPIChanges(props: OpenAPIChangesProps) {
     });
 
     return (
-
             <ConfigProvider
                 theme={{
                     algorithm: theme.darkAlgorithm,
@@ -67,16 +69,13 @@ function OpenAPIChanges(props: OpenAPIChangesProps) {
                         colorPrimary: 'rgba(98, 196, 255, 1.0)',
                         colorPrimaryActive: '#b685ff',
                     },
-                }}
-            >
+                }}>
                 <div id="main_container">
                     <section className="report-header">
                         <Header/>
                         <hr/>
                         <ReportContainer/>
-                        <hr/>
                     </section>
-
                     <section className="report-changes">
                         <Tabs
                             destroyInactiveTabPane={true}
@@ -98,7 +97,6 @@ function OpenAPIChanges(props: OpenAPIChangesProps) {
                             ]}
                         />
                     </section>
-
                 </div>
                 <footer>
                     Generated by <a href='https://pb33f.io/openapi-changes'>openapi-changes</a>&nbsp;|&nbsp;
@@ -106,8 +104,5 @@ function OpenAPIChanges(props: OpenAPIChangesProps) {
                     Developed by <a href='https://pb33f.io'>pb33f heavy industries</a>
                 </footer>
             </ConfigProvider>
-
     );
 }
-
-export default OpenAPIChanges;
