+++ html-report/ui/src/components/editor/Editor.tsx
@@ -1,12 +1,13 @@
-import React, {useContext, useEffect, useRef, useState} from "react";
+// Copyright 2023 Princess B33f Heavy Industries / Dave Shanley
+// SPDX-License-Identifier: MIT
+
+import React, {useEffect, useRef} from "react";
 import {DiffEditor, Monaco} from "@monaco-editor/react"
 import {Change} from '@/model';
 import '@/components/tree/Tree.css'
 import {EditorState, ReportState, useEditorStore, useReportStore} from "@/model/store";
 import {Switch} from "antd";
 import './Editor.css';
-import {ReportStoreContext} from "@/OpenAPIChanges";
-import {useStore} from "zustand";
 
 export interface EditorComponentProps {
     currentChange: Change | null;
@@ -18,12 +19,12 @@ const updatePosition = (currentChange: Change, editorRef: any, origCol: number, 
                         newCol: number, newLine: number) => {
     // if this is a removal
     if (origLine && !newLine) {
-            editorRef.current.setPosition({column: 1, lineNumber: 1});
-            editorRef.current.revealLinesInCenter(1, 1);
-            editorRef.current.getOriginalEditor().setPosition({column: origCol, lineNumber: origLine})
-            editorRef.current.getOriginalEditor().revealPositionInCenter({column: origCol, lineNumber: origLine})
-            editorRef.current.getOriginalEditor().revealLinesInCenter(origLine, origLine);
-            editorRef.current.getOriginalEditor().focus();
+        editorRef.current.setPosition({column: 1, lineNumber: 1});
+        editorRef.current.revealLinesInCenter(1, 1);
+        editorRef.current.getOriginalEditor().setPosition({column: origCol, lineNumber: origLine})
+        editorRef.current.getOriginalEditor().revealPositionInCenter({column: origCol, lineNumber: origLine})
+        editorRef.current.getOriginalEditor().revealLinesInCenter(origLine, origLine);
+        editorRef.current.getOriginalEditor().focus();
     } else {
         if (origLine) {
             editorRef.current.getOriginalEditor()
@@ -50,25 +51,22 @@ export function EditorComponent(props: EditorComponentProps) {
     const modifiedSpec: string | undefined = useReportStore((report: ReportState) => report.selectedReportItem?.modifiedSpec);
     const editorRef = useRef<any>(null);
     const monacoRef = useRef<any>(null);
-    const [mod] = useState(modifiedSpec);
-    const [orig] = useState(originalSpec);
     const currentChange = props.currentChange
     const sbs = props.sideBySideEditor;
-
     const setSbs = useEditorStore((editor: EditorState) => editor.setSideBySide);
 
     useEffect(() => {
         if (currentChange) {
             if (editorRef.current) {
-                    requestAnimationFrame(() => {
-                        setTimeout(()=> {
-                            updatePosition(currentChange, editorRef,
-                                currentChange.context.originalColumn,
-                                currentChange.context.originalLine,
-                                currentChange.context.newColumn,
-                                currentChange.context.newLine)
-                        }, 10)
-                    })
+                requestAnimationFrame(() => {
+                    setTimeout(() => {
+                        updatePosition(currentChange, editorRef,
+                            currentChange.context.originalColumn,
+                            currentChange.context.originalLine,
+                            currentChange.context.newColumn,
+                            currentChange.context.newLine)
+                    }, 10)
+                })
             }
         }
     });
@@ -147,8 +145,8 @@ export function EditorComponent(props: EditorComponentProps) {
     let height = props.height
     const mobile = (window.innerWidth < 1000)
     if (!height) {
-        if(mobile) {
-           height = "calc(100vh - 553px)"
+        if (mobile) {
+            height = "calc(100vh - 553px)"
         } else {
             height = "calc(100vh - 316px)"
         }
@@ -158,7 +156,6 @@ export function EditorComponent(props: EditorComponentProps) {
         setSbs(checked)
     }
 
-
     return (
         <section className='editor'>
             <span className='editor-flip-controls'>
@@ -166,14 +163,14 @@ export function EditorComponent(props: EditorComponentProps) {
                 <Switch size='small' defaultChecked onChange={flipEditor} checked={checked}/>
             </span>
             <DiffEditor
-            height={height}
-            original={orig}
-            modified={mod}
-            onMount={handleEditorDidMount}
-            theme={"pb33f"}
-            language="yaml"
-            options={options}
-        />
+                height={height}
+                original={originalSpec}
+                modified={modifiedSpec}
+                onMount={handleEditorDidMount}
+                theme={"pb33f"}
+                language="yaml"
+                options={options}
+            />
         </section>
     );
 }
