+++ html-report/ui/src/components/graph/GraphView.tsx
@@ -1,3 +1,6 @@
+// Copyright 2023 Princess B33f Heavy Industries / Dave Shanley
+// SPDX-License-Identifier: MIT
+
 import {DrawerComponent, DrawerProps} from "@/components/drawer/Drawer";
 import HorizontalFlow from "@/components/graph/Flow";
 import React from "react";
@@ -12,9 +15,9 @@ export interface GraphViewProps {
 export const GraphViewComponent = (props: GraphViewProps) => {
     const drawerOpen = useDrawerStore((state: DrawerState) => state.drawerOpen)
     return (
-        <>
-         <HorizontalFlow />
-         <DrawerComponent open={drawerOpen} />
-        </>
+        <section>
+            <HorizontalFlow/>
+            <DrawerComponent open={drawerOpen}/>
+        </section>
     )
 }
