+++ html-report/ui/src/components/navigation/TimelineChangeItemTitle.tsx
@@ -1,3 +1,6 @@
+// Copyright 2023 Princess B33f Heavy Industries / Dave Shanley
+// SPDX-License-Identifier: MIT
+
 import React from "react";
 
 export interface TimelineChangeItemTitleProps {
