+++ html-report/ui/src/model/beefy-tree-node.ts
@@ -1,3 +1,6 @@
+// Copyright 2023 Princess B33f Heavy Industries / Dave Shanley
+// SPDX-License-Identifier: MIT
+
 import {DataNode} from "antd/es/tree";
 import {Change} from "./change";
