+++ html-report/ui/src/model/change.ts
@@ -1,3 +1,6 @@
+// Copyright 2023 Princess B33f Heavy Industries / Dave Shanley
+// SPDX-License-Identifier: MIT
+
 export interface ChangeContext {
     newColumn: number;
     newLine: number;
