+++ html-report/ui/src/model/graph.ts
@@ -1,3 +1,6 @@
+// Copyright 2023 Princess B33f Heavy Industries / Dave Shanley
+// SPDX-License-Identifier: MIT
+
 import {EdgeData, NodeData} from "reaflow";
 import {BeefyTreeNode} from "@/model/beefy-tree-node";
