+++ html-report/ui/src/model/report.ts
@@ -1,3 +1,6 @@
+// Copyright 2023 Princess B33f Heavy Industries / Dave Shanley
+// SPDX-License-Identifier: MIT
+
 import {BeefyTreeNode} from "@/model/beefy-tree-node";
 import {GraphData} from "@/model/graph";
 import {ChangeStatistics} from "@/model/change";
