+++ html-report/ui/src/model/store.ts
@@ -1,3 +1,6 @@
+// Copyright 2023 Princess B33f Heavy Industries / Dave Shanley
+// SPDX-License-Identifier: MIT
+
 import create from "zustand";
 import {Change} from "./change";
 import React from "react";
